[launcher_app] treat warnings as errors

+++ src/applications/launcher_app/LauncherService.cpp
@@ -258,7 +258,7 @@ void LauncherService::ServiceMain( DWORD, LPTSTR* )
     std::vector< char* > args = boost::assign::list_of< char* >( &appName[ 0 ] )( &arg[ 0 ] );
 
     pInstance_->pFacade_.reset( new LauncherFacade( pInstance_->path_.parent_path().string() ) );
-    pInstance_->pFacade_->Initialize( args.size(), &args[ 0 ] );
+    pInstance_->pFacade_->Initialize( static_cast< int >( args.size() ), &args[ 0 ] );
     pInstance_->isRunning_ = true;
 
     MT_LOG_INFO_MSG( "ServiceMain running" );
